Filter on IDs
Users can now filter stories/epics/themes on ID by typing IDs in a
filter textbox contained in the header. This textbox can be
automatically populated if the user selects some items and presses the
new "filter selected" button, also contained in the header. In order to
display to the user what IDs different items have, a small ID text has
been added in the bottom right corner of each element.

The active filter is transferred to the web URL, making it easier to
share a filter with other users by simply copying the link.

To review this change easier, whitespace changes can be ignored by
adding ?w=1 to the URL on any github.com page that is showing a diff.

fixes sonyxperiadev/BacklogTool#25